refactor(shared): flatten features/query — drop server/ subdir
