+++ packages/shared/src/features/query/nullIfEmptyFilter.ts
