+++ packages/shared/src/server/index.ts
@@ -138,6 +138,6 @@ export * from "./deletionGuard";
 export * from "./analytics-integrations/types";
 
 // query (dashboard / monitor data model — server-only)
-export * from "../features/query/server/queryBuilder";
-export * from "../features/query/server/queryExecutor";
-export { NULL_IF_EMPTY_RE } from "../features/query/server/nullIfEmptyFilter";
+export * from "../features/query/queryBuilder";
+export * from "../features/query/queryExecutor";
+export { NULL_IF_EMPTY_RE } from "../features/query/nullIfEmptyFilter";
